+++ Editor/GameProject/OpenProject.cs
@@ -1,5 +1,4 @@
-﻿using Editor.Common;
-using Editor.Common.Enums;
+﻿using Editor.Common.Enums;
 using Editor.Utilities;
 using System.Collections.ObjectModel;
 using System.Diagnostics;