+++ Changelog.md
@@ -3,6 +3,7 @@
 ## UI 
  - Content Browser list view
  - Content Browser tile view
+ - Custom Save File Dialog
 
 # v0.3.4-beta - 2025.6.24.